authentication & autherization

+++ cards/routes/cardsController.js
@@ -6,6 +6,7 @@ import {
   updateCard,
 } from "../services/cardsService.js";
 import { auth } from "../../auth/services/authService.js";
+import { getCardByIdFromDb } from "../services/cardsDataService.js";
 
 const router = express.Router();
 
@@ -20,7 +21,12 @@ router.get("/", async (req, res) => {
 
 router.post("/", auth, async (req, res) => {
   const newCard = req.body;
-  const cardResult = await createNewCard(newCard);
+  const user = req.user;
+  if (!user.isBusiness) {
+    return res.status(403).send("Only Business user can create cards");
+  }
+  const cardResult = await createNewCard(newCard, user._id);
+
   if (cardResult) {
     res.status(201).send("New card added successfully");
   } else {
@@ -38,8 +44,18 @@ router.get("/:id", async (req, res) => {
   }
 });
 
-router.delete("/:id", async (req, res) => {
+router.delete("/:id", auth, async (req, res) => {
   const { id } = req.params;
+  const user = req.user;
+  //בדיקה אם המשתמש הוא אדמין או הבעלים של הכרטיס?
+  const card = await getCardByIdFromDb(id);
+
+  if (!user.isAdmin && card?.user_id !== user._id) {
+    return res
+      .status(403)
+      .send("Only Admin user Or owner of card can delete it");
+  }
+
   const idOfDeletedCard = await deleteCard(id);
   if (idOfDeletedCard) {
     res.send("Card deleted successfully");
@@ -48,7 +64,7 @@ router.delete("/:id", async (req, res) => {
   }
 });
 
-router.put("/:id", async (req, res) => {
+router.put("/:id", auth, async (req, res) => {
   const { id } = req.params;
   const newCard = req.body;
   const modifiedCard = await updateCard(id, newCard);
